// modules/loan/modules/third-party-loans/enum/third-party-loans-control-name.enum.ts
export enum AttributeThirdPartyLoansTable {
  FILTER = 'filter'
}

export enum ThirdPartyLoansTableKeys {

  IDENTIFIER = 'identifier',
  ALIAS = 'alias',
  ACCOUNT = 'account',
  CURRENCY_CODE = 'currencyCode',
  TYPE_ACCOUNT = 'productLabel',
  NAME = 'name',
  TYPE = 'type'
}

export enum AttributeFormThirdPartyLoansCrud {
  TYPE_ACCOUNT = 'typeAccount',
  NUMBER_ACCOUNT = 'numberAccount',
  NAME_ACCOUNT = 'nameAccount',
  ACCOUNT_STATE = 'accountState',
  AMOUNT = 'amount',
  COMMENT = 'comment',
  CURRENCY = 'CURRENCY',
  ALIAS = 'alias',
  EMAIL = 'email',
  DATE = 'datePaymentTPL',
  TOTAL_PAYMENT = "totalPayment",
  PAYMENT_OPTION = "paymentOption",
  MOUNT_PAYMENT = "mountPayment",
  QUOTAS = 'quotas',
}